requirements.txt
beir
rank_bm25
sentence-transformers
faiss-cpu
numpy
tqdm